Extract implicits from the exercise bodies
